# docs/openapi/agent-api.v1.yaml
openapi: 3.0.3
info:
  title: AI Finance - Agent API
  version: 1.0.0
  description: |
    Multi-Agent Orchestration API cho hệ thống AI Finance Assistant.
    
    **Các endpoints:**
    - POST /api/v1/agent/query - Query đồng bộ
    - POST /api/v1/agent/query/stream - Query với SSE streaming (khuyên dùng)
    - GET /api/v1/agent/tools - Danh sách tools
    - POST /api/v1/agent/mock/query - Mock data cho testing
    
    **Changelog:**
    - v1.0.0 (2026-02-12): Initial release

servers:
  - url: http://localhost:8001/api/v1
    description: Development server

paths:
  /agent/query:
    post:
      summary: Execute agent query (non-streaming)
      description: |
        Thực thi agent workflow và trả về kết quả cuối cùng.
        
        **Lưu ý:** Endpoint này đồng bộ, có thể mất 10-30 giây.
        Nên dùng `/agent/query/stream` để có trải nghiệm tốt hơn.
      operationId: agentQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentQueryRequest'
            examples:
              basic:
                summary: Basic query
                value:
                  query: "Giá cổ phiếu VCB hôm nay"
              advanced:
                summary: Advanced query with options
                value:
                  query: "So sánh VCB và VNM"
                  maxIterations: 3
                  sessionId: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Successful execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentQueryResponse'
              example:
                success: true
                query: "Giá cổ phiếu VCB hôm nay"
                answer: "VCB đang ở 95,500 VND (+1.2%)"
                iterations: 1
                executionTime: 4.52
                planSummary:
                  intent: "User wants current price of VCB"
                  toolsPlanned: ["vnstock_get_quote"]
                  canParallel: false
                toolExecutions:
                  - tool: "vnstock_get_quote"
                    success: true
                    executionTime: 0.8
                criticReviews:
                  - approved: true
                    score: 8
                    reasoning: "Good plan"
                    concerns: []
                    suggestions: []
                    requiresChanges: false
                meta:
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
                  timestamp: "2026-02-12T07:20:36.104Z"
                  duration: 4520
                  version: "v1"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /agent/query/stream:
    post:
      summary: Execute agent query with SSE streaming (KHUYÊN DÙNG)
      description: |
        Thực thi agent workflow và stream progress qua Server-Sent Events (SSE).
        
        **Event flow:**
        1. START - Workflow bắt đầu
        2. PLANNING - Orchestrator đang phân tích
        3. PLAN_CREATED - Plan được tạo
        4. CRITIC_REVIEWED - Critic đánh giá plan
        5. TOOL_EXECUTING - Đang chạy tools
        6. TOOL_COMPLETED - Tools hoàn thành
        7. SYNTHESIZING - Đang tổng hợp kết quả
        8. FINAL_RESULT - Kết quả cuối cùng
        
        **Format:**
        ```
        data: {"eventType": "START", "payload": {...}, "metadata": {...}, "progress": {...}}\n\n
        data: {"eventType": "TOOL_EXECUTING", ...}\n\n
        data: {"eventType": "FINAL_RESULT", ...}\n\n
        ```
      operationId: agentQueryStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentQueryRequest'
      responses:
        '200':
          description: SSE Stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/AgentEventStream'

  /agent/tools:
    get:
      summary: List available tools
      description: Lấy danh sách các tools mà agent có thể sử dụng
      operationId: listTools
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolsListResponse'

  /agent/mock/query:
    post:
      summary: Mock query endpoint (cho FE development)
      description: |
        Trả về mock data ngay lập tức, không gọi LLM hay external APIs.
        Dùng để FE dev/test mà không tốn quota.
      operationId: mockAgentQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentQueryRequest'
      responses:
        '200':
          description: Mock response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentQueryResponse'

  /agent/mock/query/stream:
    post:
      summary: Mock streaming endpoint (cho FE development)
      description: |
        Stream mock events với delay giả lập.
        Dùng để test UI/UX mà không cần BE chạy thật.
      operationId: mockAgentQueryStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentQueryRequest'
      responses:
        '200':
          description: Mock SSE Stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/AgentEventStream'

components:
  schemas:
    # ============================================
    # REQUEST SCHEMAS
    # ============================================
    AgentQueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          description: Câu hỏi của user
          example: "Giá cổ phiếu VCB hôm nay"
        sessionId:
          type: string
          format: uuid
          description: Session ID để theo dõi
          example: "550e8400-e29b-41d4-a716-446655440000"
        maxIterations:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
          description: Số lần loop tối đa

    # ============================================
    # RESPONSE SCHEMAS
    # ============================================
    AgentQueryResponse:
      type: object
      required:
        - success
        - query
        - answer
        - iterations
        - executionTime
        - meta
      properties:
        success:
          type: boolean
        query:
          type: string
        answer:
          type: string
          description: Câu trả lờ cuối cùng
        iterations:
          type: integer
          description: Số vòng lặp đã thực hiện
        executionTime:
          type: number
          format: float
          description: Thờ gian thực thi (giây)
        planSummary:
          $ref: '#/components/schemas/PlanSummary'
        toolExecutions:
          type: array
          items:
            $ref: '#/components/schemas/ToolExecution'
        criticReviews:
          type: array
          items:
            $ref: '#/components/schemas/CriticReview'
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    PlanSummary:
      type: object
      properties:
        intent:
          type: string
          description: Intent được AI hiểu
        toolsPlanned:
          type: array
          items:
            type: string
          example: ["tavily_web_search", "vnstock_get_quote"]
        canParallel:
          type: boolean

    ToolExecution:
      type: object
      properties:
        tool:
          type: string
        success:
          type: boolean
        executionTime:
          type: number

    CriticReview:
      type: object
      properties:
        approved:
          type: boolean
        score:
          type: number
          minimum: 0
          maximum: 10
        reasoning:
          type: string
        concerns:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        requiresChanges:
          type: boolean
        iterationFeedback:
          type: string
          nullable: true

    ResponseMeta:
      type: object
      required:
        - requestId
        - timestamp
        - duration
        - version
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        duration:
          type: integer
          description: Thờ gian xử lý (ms)
        version:
          type: string
          example: "v1"

    ToolsListResponse:
      type: object
      properties:
        success:
          type: boolean
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolInfo'

    ToolInfo:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          type: object
          additionalProperties: true
        useCases:
          type: array
          items:
            type: string

    # ============================================
    # SSE EVENT SCHEMAS
    # ============================================
    AgentEventStream:
      oneOf:
        - $ref: '#/components/schemas/StartEvent'
        - $ref: '#/components/schemas/PlanningEvent'
        - $ref: '#/components/schemas/PlanCreatedEvent'
        - $ref: '#/components/schemas/CriticReviewedEvent'
        - $ref: '#/components/schemas/ToolExecutingEvent'
        - $ref: '#/components/schemas/ToolCompletedEvent'
        - $ref: '#/components/schemas/SynthesizingEvent'
        - $ref: '#/components/schemas/FinalResultEvent'
        - $ref: '#/components/schemas/ErrorEvent'

    # Base Event
    AgentEventBase:
      type: object
      required:
        - eventType
        - metadata
        - progress
      properties:
        eventType:
          type: string
          enum:
            - START
            - PLANNING
            - PLAN_CREATED
            - CRITIC_REVIEWED
            - TOOL_EXECUTING
            - TOOL_COMPLETED
            - SYNTHESIZING
            - FINAL_RESULT
            - ERROR
        metadata:
          type: object
          required:
            - iteration
            - timestamp
            - executionTime
          properties:
            iteration:
              type: integer
            timestamp:
              type: string
              format: date-time
            executionTime:
              type: number
              description: Tổng thờ gian đã chạy (ms)
        progress:
          type: object
          required:
            - currentStep
            - totalSteps
            - percentComplete
          properties:
            currentStep:
              type: integer
            totalSteps:
              type: integer
            percentComplete:
              type: integer
              minimum: 0
              maximum: 100

    StartEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [START]
            payload:
              type: object
              required:
                - query
              properties:
                query:
                  type: string

    PlanningEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [PLANNING]
            payload:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  example: "Orchestrator đang phân tích..."

    PlanCreatedEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [PLAN_CREATED]
            payload:
              type: object
              required:
                - plan
              properties:
                plan:
                  $ref: '#/components/schemas/OrchestratorPlan'

    OrchestratorPlan:
      type: object
      required:
        - originalQuery
        - interpretedIntent
        - tools
        - canParallel
        - expectedOutcome
      properties:
        originalQuery:
          type: string
        interpretedIntent:
          type: string
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolPlan'
        canParallel:
          type: boolean
        expectedOutcome:
          type: string

    ToolPlan:
      type: object
      required:
        - toolName
        - parameters
        - reason
        - priority
      properties:
        toolName:
          type: string
        parameters:
          type: object
          additionalProperties: true
        reason:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 10

    CriticReviewedEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [CRITIC_REVIEWED]
            payload:
              $ref: '#/components/schemas/CriticReview'

    ToolExecutingEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [TOOL_EXECUTING]
            payload:
              type: object
              required:
                - toolName
                - parameters
              properties:
                toolName:
                  type: string
                parameters:
                  type: object

    ToolCompletedEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [TOOL_COMPLETED]
            payload:
              type: object
              required:
                - toolName
                - success
                - executionTime
              properties:
                toolName:
                  type: string
                success:
                  type: boolean
                executionTime:
                  type: number
                result:
                  type: string

    SynthesizingEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [SYNTHESIZING]
            payload:
              type: object
              required:
                - message
              properties:
                message:
                  type: string

    FinalResultEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [FINAL_RESULT]
            payload:
              $ref: '#/components/schemas/AgentQueryResponse'

    ErrorEvent:
      allOf:
        - $ref: '#/components/schemas/AgentEventBase'
        - type: object
          properties:
            eventType:
              type: string
              enum: [ERROR]
            payload:
              type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: object

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INVALID_QUERY
              message: "Query cannot be empty"
            meta:
              requestId: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2026-02-12T07:20:36.104Z"
              duration: 12
              version: "v1"

    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        meta:
          $ref: '#/components/schemas/ResponseMeta'